substrateinterface/base.py: init_runtime() improvement
If init_runtime(), or any of its parts, raises an exception, various
fields are already set. A retry of init_runtime() then fails to e.g. set
self.metadata, as these fields are used as an indication of being
already initialized.

self.metadata == None then leads to exceptions at various places.

This patch fixes this by setting these fields on self only on successful
initialization, which is on reaching the end of init_runtime().

One problem is that init_runtime() calls self.get_constant() which calls
init_runtime(). This would lead to an infinite loop if no re-entrancy
guard is present. Therefore self.block_hash is set before calling
self.get_constant(), but on failure it is reset to its original value.